Actually rename the directory correctly.
