+++ cmd/cosign/cli/generate.go
