+++ cmd/cosign/cli/sign_blob.go
