+++ cmd/cosign/cli/verify_blob.go
